new iterator for the rb tree
